requirements.txt
appdirs==1.4.4
arrow==1.1.0
asgiref==3.3.4
binaryornot==0.4.4
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
colorama==0.4.4
cookiecutter==1.7.3
crispy-tailwind==0.5.0
distlib==0.3.2
dj-database-url==0.5.0
dj-email-url==1.0.2
Django==3.2.3
django-cache-url==3.2.3
django-crispy-forms==1.12.0
django-extensions==3.1.3
django-phonenumber-field==5.2.0
django-tailwind==2.0.1
environs==8.0.0
filelock==3.0.12
graphviz==0.16
gunicorn==20.0.4
idna==2.10
importlib-metadata==1.7.0
Jinja2==3.0.1
jinja2-time==0.2.0
MarkupSafe==2.0.1
marshmallow==3.12.1
phonenumbers==8.12.25
Pillow==8.2.0
poyo==0.5.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pydot==1.4.2
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.17.1
python-slugify==5.0.2
pytz==2021.1
qrcode==6.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
urllib3==1.26.5
virtualenv==20.0.31
whitenoise==5.1.0
zipp==3.4.1
django-cors-headers==*
django-intl-tel-input2
django_multivalueformfield==*
django-multiupload==*
django-celery-beat===*